+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Move group affiliation of attendees and lecturers to the `courseGroups` attribute to have it encapsulated there
+
 ## v0.2.1
 
 - for a start only cache one semster back and ahead to not overcrowd existing course selectors which don't filter by semester yet